% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.R
\name{auto_handler}
\alias{auto_handler}
\title{Handle Object Types Automatically}
\usage{
auto_handler(use_plotly_instead_of_ggplot = TRUE)
}
\arguments{
\item{use_plotly_instead_of_ggplot}{Use plot.ly to render ggplot2 charts. It is \code{TRUE} by default.}
}
\description{
Handle Object Types Automatically
}
