% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{configure}
\alias{configure}
\title{Configure 'dstack'}
\usage{
configure(
  profile = "default",
  user,
  token,
  persist = c("global", "local", "in_place"),
  server = NULL,
  dstack_dir = .dstack_env$dstack_dir
)
}
\arguments{
\item{profile}{A name of profile. It will be "default" if not specified.}

\item{user}{Username in 'dstack'.}

\item{token}{A token. It can be obtained from \url{https://dstack.ai} web site.}

\item{persist}{Persistence level. It can be 'local' - this means that config will be stored in working directory,
'global' - config will be stored in user's home,cor 'in-place' - in this case config will be store in the memory
and exists only while R session exists.}

\item{server}{Server to connect. By default it's \code{NULL}, so default API endpoint will be used.}

\item{dstack_dir}{Directory where dstack files are stored. By default it is '.dstack', so in the case of global persistence
path to config file will be \code{~/.dstack/config.yaml}.}
}
\description{
Function allows to add or replace profile.
}
