% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{dataframe_handler}
\alias{dataframe_handler}
\title{Handle 'data.frame' Objects}
\usage{
dataframe_handler(col_names = TRUE, row_names = FALSE)
}
\arguments{
\item{col_names}{Save column names. \code{TRUE} by default.}

\item{row_names}{Save row names. \code{FALSE} by default.}
}
\description{
Handle 'data.frame' Objects
}
