% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable.R
\name{datatable_handler}
\alias{datatable_handler}
\title{Handle 'data.table' Objects}
\usage{
datatable_handler(col_names = TRUE, row_names = FALSE)
}
\arguments{
\item{col_names}{Save column names. \code{TRUE} by default.}

\item{row_names}{Save row names. \code{FALSE} by default.}
}
\description{
Handle 'data.table' Objects
}
