% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{ggplot_handler}
\alias{ggplot_handler}
\title{Handles 'ggplot2' Objects}
\usage{
ggplot_handler(dpi = 300, width = NA, height = NA, format = "png")
}
\arguments{
\item{dpi}{DPI, default is 300.}

\item{width}{Image width.}

\item{height}{Image height.}

\item{format}{Image format to use, can be "png" or "svg", by default PNG will be used.}
}
\value{
A list which contains conversion function and format inself.
}
\description{
Returns a function that converts 'ggplot2' plots to appropriate format and format itself. PNG and SVG are supported.
}
