% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{in_place_config}
\alias{in_place_config}
\title{In-Place Configuration}
\usage{
in_place_config()
}
\description{
It is used to configure dstack in R session without creating any configuration file on disk.
}
