% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.R
\name{is.datatable}
\alias{is.datatable}
\title{Check That Specified Object is 'data.table' Object}
\usage{
is.datatable(x)
}
\arguments{
\item{x}{An object to check.}
}
\description{
Check That Specified Object is 'data.table' Object
}
