% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.R
\name{is.ggplot2}
\alias{is.ggplot2}
\title{Check That Specified Object is 'ggplot2' Chart}
\usage{
is.ggplot2(x)
}
\arguments{
\item{x}{An object to check.}
}
\description{
Check That Specified Object is 'ggplot2' Chart
}
