% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.R
\name{is.plotly}
\alias{is.plotly}
\title{Check That Specified Object is 'plot.ly' Chart}
\usage{
is.plotly(x)
}
\arguments{
\item{x}{An object to check.}
}
\description{
Check That Specified Object is 'plot.ly' Chart
}
