% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{json_protocol}
\alias{json_protocol}
\title{JSON Protocol Implementation to Connect API Server}
\usage{
json_protocol(server, error = .error)
}
\arguments{
\item{server}{A server to connect.}

\item{error}{An error handling function.}
}
\value{
A function that implements JSON protocol.
}
\description{
Protocol is an abstraction which allows to send data to server.
This function implements JSON-based protocol. It provides token in
'Authorization' header.
}
