% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{push}
\alias{push}
\title{Push All Commits to Server}
\usage{
push(frame, message = NULL)
}
\arguments{
\item{frame}{A frame to push.}

\item{message}{Push message. \code{NULL} by default.}
}
\value{
Stack URL.
}
\description{
Tis function is used to send a banch of commited plots to server or say server that operation with this frame is done.
In the case of 'auto_push' mode it sends only a total number
of views in the frame. So call this method is obligatory to close the frame anyway.
}
\examples{
\donttest{
library(ggplot2)
library(dstack)
image <- qplot(clarity, data = diamonds, fill = cut, geom = "bar")
frame <- create_frame(stack = "diamonds")
frame <- commit(frame, image, "Diamonds bar chart")
print(push(frame)) # print actual stack URL
}
}
