% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{use_config}
\alias{use_config}
\title{Use Specified Configuration}
\usage{
use_config(config_func)
}
\arguments{
\item{config_func}{A function to be used for configuration. It can be \code{yaml_config} or \code{in_place_config}.}
}
\description{
Use Specified Configuration
}
\examples{
use_config(yaml_config) # to use standard YAML configuration file
use_config(in_place_config) # to use "in-place" configuration which is stored in memory
}
